fix/separate product details sections into widgets
